pyproject.toml
[project]
name = "researcher-agent"
version = "0.1.0"
description = "IT-Sicherheitsarchitekt Researcher Agent mit Brave Search und Microsoft Learn MCP"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.68",
    "httpx>=0.27",
    "jinja2>=3.1",
    "typer>=0.12",
    "python-slugify>=8.0",
    "python-dotenv>=1.0",
    "markdown-it-py>=3.0",
]

[project.scripts]
researcher = "researcher.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["researcher"]

[tool.hatch.build.targets.wheel.force-include]
"researcher/templates" = "researcher/templates"
"researcher/static" = "researcher/static"
"researcher/prompts" = "researcher/prompts"
